+++ changelog.md
@@ -9,6 +9,7 @@
 * [EGD-3138] Add handling for speed dial number duplicate window.
 * [EGD-3275] Add confirmation windows after deleting/blocking contact.
 * [EGD-3261] Empty screen at the top fix in SMS thread view 
+* [EGD-3278] Fix ListView on page refresh item order after page change. 
 
 ## **04-06-2020**
 * Initial changelog file created